% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.adwords.keyword.r
\name{as.adwords.keyword}
\alias{as.adwords.keyword}
\title{As AdWords Keyword}
\usage{
as.adwords.keyword(keyword_vector)
}
\arguments{
\item{keyword_vectoro}{A vector of keywords from AdWords}
}
\description{
This function formats AdWords keywords data by removing special characters and ensuring proper typing.
}
\examples{
as.device(device_vector)
as.device(google_adwords_report$device)
}

