% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.week.r
\name{as.week}
\alias{as.week}
\title{As Week}
\usage{
as.week(date_vector)
}
\arguments{
\item{date_vector}{A vector of dates.}
}
\description{
This function converts a vector of date objects 
(whose day may be any time within Sunday to Monday)
of the week, into the date for the Monday of that week.
}
\examples{
date_vector <- (c(Sys.Date()))
date_vector
[1] "2016-05-07" #Thursday
as.week(date_vector)
[1] "2016-05-02" # Preceeding Monday of the same week

date_vector <- (c(Sys.Date()))
date_vector
[1] "2016-05-01" # Sunday
as.week(date_vector)
[1] "2016-05-02" # Following monday of the same week
}

