% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_lookup_table.r
\name{campaign_lookup_table}
\alias{campaign_lookup_table}
\title{Create Campaign Lookup Table}
\usage{
campaign_lookup_table(from = Sys.Date(), to = Sys.Date())
}
\arguments{
\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}
}
\value{
A data frame with a lookup table of campaign IDs and campaign names.
}
\description{
Creates a lookup table for all campaign names that exist within a date range.  
That is, creates one column for campaign name, and another column with campaign ID.
}
\examples{
campaign_lookup_table(from=20150101, to=20151231)
}

