% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_through_rate_over_time.r
\name{click_through_rate_over_time}
\alias{click_through_rate_over_time}
\title{Create plot of AdWords click through rate over time}
\usage{
click_through_rate_over_time(keywords_elog, plot = TRUE)
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}

\item{plot}{logical. If TRUE, creates a plot of the data.  If false, creates no plot, Defaults to \code{TRUE}.}
}
\value{
The data frame used to create a plot of clickthrough rate over time.
}
\description{
Takes in a AdWords event log and processes click through rate over time.
}
\examples{
overall_performance_over_time(elog_data_frame)
}

