% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_per_click_report.r
\name{contribution_per_click_report}
\alias{contribution_per_click_report}
\title{Contribution per click report}
\usage{
contribution_per_click_report(keywords_elog, keyword_filter = NULL,
  campaign_filter = NULL, write_csv = TRUE,
  file = "adwords_contribution_per_click_report.csv")
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}

\item{keyword_filter}{Optional. (e.g. "paleo meals") If provided, generates a data frame and plot only for the *exact* keyword provided. Default to NULL.}

\item{campaign_filter}{Optional. (e.g. "Paleo Performers") If provided, generates a data frame and plot for campaigns where the string is included in the campaign name. Default to NULL.}

\item{write_csv}{logical. If TRUE, creates a csv of the data frame in the working directory with the filename. If false, creates no file. Defaults to \code{TRUE}.}

\item{file}{String. The filename for the generated report if \code{write_csv} is set to \code{TRUE}. Defaults to "bing_contribution_per_click_report.csv".}
}
\value{
The data frame used to create the report.
}
\description{
Take in a bing keyword event log and creates a contribution per click report
by keyword and campaign_name
}
\examples{
contribution_per_click_report(keywords_elog, keyword_filter="paleo meals", campaign_filter="Paleo Performers")
}

