% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desktop_gdn_performance_over_time.r
\name{desktop_gdn_performance_over_time}
\alias{desktop_gdn_performance_over_time}
\title{Create plot of Google Display Network mobile performance over time}
\usage{
desktop_gdn_performance_over_time(gdn_elog, campaign_filter = NULL,
  plot = TRUE)
}
\arguments{
\item{gdn_elog}{A data frame containing keywords and transaction event log data.}

\item{campaign_filter}{Optional. (e.g. "Paleo Performers") If provided, generates a data frame and plot for campaigns where the string is included in the campaign name. Default to NULL.}

\item{plot}{logical. If TRUE, creates a plot of the data.  If false, creates no plot, Defaults to \code{TRUE}.}
}
\value{
The data frame used to create a plot of mobile performance over time.
}
\description{
Takes in a AdWords event log and processes performance on desktop devices over time.
}
\examples{
overall_performance_over_time(elog_data_frame)
}

