% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_performance_over_time.r
\name{mobile_performance_over_time}
\alias{mobile_performance_over_time}
\title{Create plot of AdWords mobile performance over time}
\usage{
mobile_performance_over_time(keywords_elog, keyword_filter = NULL,
  campaign_filter = NULL, plot = TRUE)
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}

\item{keyword_filter}{Optional. (e.g. "paleo meals") If provided, generates a data frame and plot only for the *exact* keyword provided. Default to NULL.}

\item{campaign_filter}{Optional. (e.g. "Paleo Performers") If provided, generates a data frame and plot for campaigns where the string is included in the campaign name. Default to NULL.}

\item{plot}{logical. If TRUE, creates a plot of the data.  If false, creates no plot, Defaults to \code{TRUE}.}
}
\value{
The data frame used to create a plot of mobile performance over time.
}
\description{
Takes in a AdWords event log and processes performance on mobile devices over time.
}
\examples{
overall_performance_over_time(elog_data_frame)
}

