% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_campaign_performance_data.r
\name{raw_campaign_performance_data}
\alias{raw_campaign_performance_data}
\title{Get Raw Campaign Performance Data}
\usage{
raw_campaign_performance_data(from = Sys.Date(), to = Sys.Date())
}
\arguments{
\item{from}{Date object for start date in format <yyyymmdd>. Inclusive.  If left blank, defaults to Sys.Date()}

\item{to}{Date object for end date in format <yyyymmdd>. Inclusive. If left blank, defaults to Sys.Date()}
}
\value{
A data frame with the campaign performance data in the timeframe.
}
\description{
This is a utility function for that pulls a Campaign Performance Report
with the data needed for our analysis. If a field is missing that is 
supported by the \code{campaign performance report}, it should be added to
this function.
NOTE: Requires you to enter Google Authentication info or have
.google.auth.RData in your working directory.
}
\examples{
campaign_performance_data(from=20151216, to=20151219)
}

