% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_performance_over_time.r
\name{overall_performance_over_time}
\alias{overall_performance_over_time}
\title{Create plot of AdWords performance over time}
\usage{
overall_performance_over_time(keywords_elog, plot = TRUE)
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}
}
\value{
plot logical. If TRUE, creates a plot of the data.  If false, creates no plot, Defaults to \code{TRUE}.

The data frame used to create a plot of performance over time.
}
\description{
Takes in a Bing event log and processes all summary data
}
\examples{
summarize_bing_elog(elog_data_frame)
}

